saving works, but this is wip:
- updateNodeNames wipes stuff out when it is called on node add and remove
- we need to take out updateNodeNames
- w need to replace with a naming function on add
