+++ src/plugins/dataflow/model/dataflow-program-model.ts
@@ -122,6 +122,7 @@ export const DataflowProgramModel = types.
   .actions(self => ({
     // This isn't great but it is how the unique node names have been working
     updateNodeNames(){
+      console.log("| updateNodeNames called!");
       let idx = 1;
       self.nodes.forEach((node) => {
         const nodeType = NodeTypes.find( (n: NodeType) => n.name === node.name);
