+++ src/plugins/dataflow/nodes/editable-node-name.tsx
@@ -1,0 +1,30 @@
+import React, { useState } from 'react';
+import { IBaseNode } from './base-node';
+
+interface EditableNodeNameProps {
+  node:IBaseNode;
+}
+
+export const EditableNodeName: React.FC<EditableNodeNameProps> = ({ node }) => {
+  const [nodeName, setNodeName] = useState(node.model.orderedDisplayName);
+
+  function handleInputChange(e: React.ChangeEvent<HTMLInputElement>) {
+    setNodeName(e.target.value);
+  }
+
+  function saveNodeName(){
+    nodeName && node.model.setOrderedDisplayName(nodeName);
+  }
+
+  return (
+    <div className="node-name">
+      <input
+        className="node-name-input"
+        value={nodeName}
+        onChange={handleInputChange}
+        onBlur={saveNodeName}
+        disabled={node.readOnly}
+      />
+    </div>
+  );
+};
